requirements.txt
pyyaml==6.0.2
opencv-python==4.10.0.84
numpy==1.26.4
onnxruntime==1.19.2
pyopencl==2024.2.7
aiohttp==3.10.10
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
psutil==6.1.0
pyserial==3.5
smbus2==0.4.3
pytest==8.3.3
pytest-mock==3.14.0